+++ MANIFEST.in
@@ -1,7 +1,7 @@
 # Manifest syntax https://docs.python.org/2/distutils/sourcedist.html
 graft wheelhouse
 
-recursive-include repo_stats *.py
+recursive-include repostats *.py
 recursive-exclude __pycache__  *.pyc *.pyo *.orig
 
 # Include the README
