+++ pyproject.toml
@@ -11,7 +11,7 @@ exclude = "(.eggs|.git|.hg|.mypy_cache|.venv|_build|buck-out|build|dist)"
 
 [tool.isort]
 known_first_party = [
-    "repo_stats",
+    "repostats",
 ]
 skip_glob = []
 profile = "black"
