+++ repostats/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.0"
+__version__ = "0.1.7"
 __author__ = "Jiri Borovec"
 __author_email__ = "j.borovec@...cz"
 __license__ = "MIT"
